% Generated by roxygen2 (4.0.1): do not edit by hand
\name{K12}
\alias{K12}
\title{Monte Carlo Estimation of Sensitivity Indices for Models with Dependent Variables (Gaussian Copula Scheme by Kucherenko et al. 2012)}
\usage{
K12(data = NULL, distributions = NULL, parameters = NULL, N = 10^4,
  sigma = NULL, model, k = NULL, mu = NULL, threshold = c(0.01, 0.05,
  0.1), ...)
}
\arguments{
\item{data}{\code{NULL} (default) A matrix with columns taken as model input variables that follow arbitrary marginal distributions (specified by \code{distributions}), and rows as data observations. \code{data} is used only to extract variable means, and variable correlation for transformation to MVN simulated data usign copula based methods. \code{data} is not entered when providing simulation parameters \code{k,sigma,mu} for variables distributions instead.}

\item{distributions}{A list of distribution names from the list c("unif", "norm", "lnorm", "gamma"), the length of \code{ncol(data)} or \code{k} (in the case of a simulation). Default is all variables are assigned a uniform marginal distribution.}

\item{k}{\code{NULL} (default) The number of marginal distributions, specified when no data is given, and when \code{sigma} and \code{mu} are provided (such as in a simulation).}

\item{sigma}{\code{NULL} (default) The marginal distributions' covariance matrix, specified when no data is given, and when \code{k} and \code{mu} are provided (such as in a simulation). A covariance matrix, instead of a correlation matrix, is required for arbitrary-MVN distribution mapping (see \code{L86} function).}

\item{mu}{\code{NULL} (default) The marginal distributions' mean vector (of length \code{k}), specified when no data is given, and when \code{k} and \code{sigma} are provided (such as in a simulation).}

\item{model}{A function passed into \code{K12}. The model must take input variables by name/column number matching the name/column order in \code{data}, or in the order relative to parameters \code{distributions, parameters, sigma, mu} in the case of a simulation.}

\item{parameters}{\code{NULL} (default) A list containing \code{k} vectors of length two for marginal distribution fit parameters (e.g. \code{mean,sd} for the normal distribution). Parameters are calculated by defualt wtihin \code{K12} using \code{ParamFit} (maximum likelihood fit using \code{fitdistr} from the \code{MASS} package) when \code{data} is supplied. In the case of a simulation, a \code{parameters} list must be supplied.}

\item{N}{Sample size, by default \emph{=10^4}.}

\item{threshold}{A vector of any length of desired (arbitrary) sensitivity index thresholds (for return of a \code{$results} matrix indicating variables whose first order and total SI meet the threshold value. By default, the threshold is set to 0.01,0.05, and 0.10 (1\%, 5\%, and 10\% of total output variance).}

\item{...}{Additional arguments, such as those to be passed to the \code{model} function.}
}
\value{
list \code{K12} returns a list containing the first order SI estimates \code{$SY}; the total SI estimates \code{$STy}; and a printed matrix \code{$results} of first order and total SI achieving threshold values.
}
\description{
The \code{K12} function calculates variance decomposition based first order (individual) and interaction (total) sensitivity indices (SI) for models with dependent input variables (data or parameters) using Monte Carlo estimation methods, and a Gaussian copula based approach for sampling from arbitrary multivariate distributions.
}
\details{
The \code{K12} method is a generalization of the method used to compute Sobol' SI, wherein SI respond to the relative impact of interactions and input variable correlations on model output, instead of assuming variable independence (as in previous Sobol' index estimation methods). \code{K12} relies on specification of a parametric marginal distribution form to each model input variable. Marginal distributions may be any combination of the following distributions: uniform, normal, lognormal, or gamma. Simple random sampling is used for Monte Carlo estimation of SI, and a sample size of no less than \emph{N=10^4} (default) is recommended based on analytical tests. Any number of input variables may be used, however using the recommended sample size with \emph{k>10} input variables may be slow; the function estimates \emph{2k} indices (for first order and total SI) at a cost of \emph{N*(2k+1)} model evaluations. \emph{(Future iterations will implement sampling methods that require fewer samples to achieve SI estimate convergence according to analytical tests.)}
}
\examples{
\dontrun{
# Data (trees) Example
cor(trees) # dependent variables
mod_trees <- function(data) data[ ,1]*data[ ,2]*data[ ,3]
s_trees <- K12(data = trees, distributions = c("gamma", "gamma", "norm"), model = mod_trees)
s_trees$SY
s_trees$STy
s_trees$results

# Simulation Example
N <- 10^4 # observations
k <- 5 # variables
sigma <- diag(k) # independent
sim <- matrix(runif(N*k, min=0, max=10), ncol=k)
mod_sim <- function(data) (data[ ,1]*data[ ,2]) + data[ ,3] + data[ ,4] + data[ ,5]
s_sim <- K12(data = sim, model = mod_sim) # distributions = uniforms by default
s_sim$SY
s_sim$STy
s_sim$results
}
}
\references{
Kucherenko, S., S. Tarantola, and P. Annoni. "Estimation of Global Sensitivity Indices for Models with Dependent Variables." Computer Physics Communications 183, no. 4 (April 2012): 937-946. doi:10.1016/j.cpc.2011.12.020.

Saltelli, Andrea, Paola Annoni, Ivano Azzini, Francesca Campolongo, Marco Ratto, and Stefano Tarantola. "Variance Based Sensitivity Analysis of Model Output. Design and Estimator for the Total Sensitivity Index." Computer Physics Communications 181, no. 2 (February 2010): 259-270. doi:10.1016/j.cpc.2009.09.018.

Saltelli, Andrea. "Making Best Use of Model Evaluations to Compute Sensitivity Indices." Computer Physics Communications 145, no. 2 (May 15, 2002): 280-297. doi:10.1016/S0010-4655(02)00280-1.

Sobol', I.M. "Global Sensitivity Indices for Nonlinear Mathematical Models and Their Monte Carlo Estimates." Mathematics and Computers in Simulation 55, no. 1-3 (February 15, 2001): 271-280. doi:10.1016/S0378-4754(00)00270-6.
}

